segregated code for readability, in preparation for eliminating pod registry implementation
